chore: refactor source cards to squad cards
