+++ packages/shared/src/components/sources/index.ts
@@ -1,1 +1,1 @@
-export * from './SourceCard';
+export * from './SourceActions';
