+++ packages/shared/src/components/squads/index.ts
@@ -1,2 +1,3 @@
 export * from './SquadsDirectoryHeader';
 export * from './YourSquadItem';
+export * from './cards/directory/SquadGrid';
